+++ crates/web/src/handlers/lumbermill.rs
@@ -156,7 +156,7 @@ impl WoodType {
 struct PlayerRow {
     pub location: String,
     pub energy: f64,
-    pub credits: i64,
+    pub credits: i32,
     pub class: String,
     pub race: String,
 }
@@ -364,7 +364,7 @@ pub async fn lumbermill_plan_buy(
         return error_page(&app, &ctx, "Twoja umiejętność ciesielstwa jest zbyt niska.");
     }
 
-    if player_row.credits < i64::from(plan.cost) {
+    if player_row.credits < plan.cost {
         return error_page(&app, &ctx, "Nie masz wystarczająco złota.");
     }
